# backend/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ContractResponseDto": {
        "properties": {
          "billingDay": {
            "example": 10,
            "maximum": 28,
            "minimum": 1,
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "durationInMonths": {
            "example": 12,
            "maximum": 600,
            "minimum": 1,
            "type": "number"
          },
          "endDate": {
            "example": "2027-06-30",
            "format": "date",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "isRenewable": {
            "type": "boolean"
          },
          "monthlyBaseValueCents": {
            "example": 150000,
            "maximum": 2147483647,
            "minimum": 1,
            "type": "number"
          },
          "moveInDate": {
            "example": "2026-07-01",
            "format": "date",
            "type": "string"
          },
          "propertyUnitId": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ContractStatus"
              }
            ]
          },
          "tenantId": {
            "format": "uuid",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenantId",
          "propertyUnitId",
          "moveInDate",
          "endDate",
          "monthlyBaseValueCents",
          "durationInMonths",
          "billingDay",
          "isRenewable",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
      "ContractStatus": {
        "enum": [
          "ACTIVE",
          "EXPIRED",
          "TERMINATED"
        ],
        "type": "string"
      },
      "CreateContractDto": {
        "properties": {
          "billingDay": {
            "example": 10,
            "maximum": 28,
            "minimum": 1,
            "type": "number"
          },
          "durationInMonths": {
            "example": 12,
            "maximum": 600,
            "minimum": 1,
            "type": "number"
          },
          "isRenewable": {
            "type": "boolean"
          },
          "monthlyBaseValueCents": {
            "example": 150000,
            "maximum": 2147483647,
            "minimum": 1,
            "type": "number"
          },
          "moveInDate": {
            "example": "2026-07-01",
            "format": "date",
            "type": "string"
          },
          "propertyUnitId": {
            "format": "uuid",
            "type": "string"
          },
          "tenantId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "tenantId",
          "propertyUnitId",
          "moveInDate",
          "monthlyBaseValueCents",
          "durationInMonths",
          "isRenewable"
        ],
        "type": "object"
      },
...
      "PageMetaDto": {
        "properties": {
          "limit": {
            "example": 20,
            "maximum": 100,
            "minimum": 1,
            "type": "number"
          },
          "page": {
            "example": 1,
            "minimum": 1,
            "type": "number"
          },
          "total": {
            "example": 42,
            "minimum": 0,
            "type": "number"
          },
          "totalPages": {
            "example": 3,
            "minimum": 0,
            "type": "number"
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "totalPages"
        ],
        "type": "object"
      },
      "PaginatedContractsResponseDto": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ContractResponseDto"
            },
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
...
      "ProblemDetailsDto": {
        "properties": {
          "detail": {
            "example": "A requisição contém dados inválidos.",
            "type": "string"
          },
          "errors": {
            "description": "Demais mensagens de validação, quando houver mais de uma.",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "instance": {
            "example": "/contracts?page=0",
            "type": "string"
          },
          "requestId": {
            "example": "0f2a9a56-b0d5-4ac5-8f1e-791e9fc92347",
            "nullable": true,
            "type": "string"
          },
          "status": {
            "example": 400,
            "maximum": 599,
            "minimum": 400,
            "type": "number"
          },
          "timestamp": {
            "example": "2026-07-12T18:30:00.000Z",
            "format": "date-time",
            "type": "string"
          },
          "title": {
            "example": "BadRequestException",
            "type": "string"
          },
          "type": {
            "example": "https://tenancy-ledger.local/problems/http-400",
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "requestId",
          "timestamp"
        ],
        "type": "object"
      },
...
      "RenewContractDto": {
        "properties": {
          "extraMonths": {
            "example": 12,
            "maximum": 600,
            "minimum": 1,
            "type": "number"
          }
        },
        "required": [
          "extraMonths"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "description": "JWT de acesso.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "API para gestão auditável de inquilinos, imóveis, contratos e cobranças.",
    "title": "Tenancy Ledger API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/contracts": {
      "get": {
        "operationId": "ContractController_list",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ContractStatus"
            }
          },
          {
            "in": "query",
            "name": "tenantId",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "propertyUnitId",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedContractsResponseDto"
                }
              }
            },
            "description": "",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Requisição inválida.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Token ausente, inválido ou expirado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Papel sem permissão para a operação.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Limite de requisições excedido.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Erro interno inesperado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar contratos",
        "tags": [
          "Contratos"
        ]
      },
      "post": {
        "operationId": "ContractController_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContractDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractResponseDto"
                }
              }
            },
            "description": "",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Requisição inválida.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Token ausente, inválido ou expirado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Papel sem permissão para a operação.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Locatário ou imóvel não encontrado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "A unidade possui contrato com vigência sobreposta.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Regra de negócio ou valor de domínio inválido.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Limite de requisições excedido.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Erro interno inesperado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar contrato",
        "tags": [
          "Contratos"
        ]
      }
    },
    "/contracts/{id}": {
      "get": {
        "operationId": "ContractController_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractResponseDto"
                }
              }
            },
            "description": "",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Requisição inválida.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Token ausente, inválido ou expirado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Papel sem permissão para a operação.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Contrato não encontrado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Limite de requisições excedido.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Erro interno inesperado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Consultar contrato",
        "tags": [
          "Contratos"
        ]
      }
    },
    "/contracts/{id}/renew": {
      "patch": {
        "operationId": "ContractController_renew",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenewContractDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractResponseDto"
                }
              }
            },
            "description": "",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Requisição inválida.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Token ausente, inválido ou expirado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Papel sem permissão para a operação.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Contrato não encontrado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Contrato não renovável ou renovação sobreposta.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Regra de negócio ou valor de domínio inválido.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Limite de requisições excedido.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Erro interno inesperado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Renovar contrato",
        "tags": [
          "Contratos"
        ]
      }
    },
...
  },
  "servers": [],
...
}